ai-core-summary-chatbot/proxy/requirements.txt
pypdf==3.16.3
pdfminer.six==20221105
streamlit==1.27.2
pip==23.2.1